Fix: tc_almacen_usados_resumen nunca funciono (42804 escondido tras un catch)
El 400 de produccion era "structure of query does not match function result
type": la funcion declara unidades_stock numeric pero movimientos_stock.
cantidad es entera, sum() devuelve bigint y RETURN QUERY exige el tipo
exacto. tc_stock_almacen_empresa esquiva esto con casts ::numeric (por eso
el stock de panel y APK funciona); a esta se le olvidaron, y el
.catch(() => null) de la pantalla lo tapo desde el primer dia — las
tarjetas y el aviso de regularizacion simplemente no salian.

Reproducido en local con cantidad entera (mismo error literal) y verificado
el arreglo: fichas=14, apuntes=3, los numeros reales de produccion. Se
corrige tambien la version del fichero original para que una reejecucion
no reinstale el fallo.

+++ supabase/migrations/tyrecontrol_almacen_usados.sql
@@ -178,7 +178,10 @@ begin
     (select count(*) from tc_neumaticos n
       where n.empresa_id = p_empresa and n.activo and n.estado in ('almacen','reservado')
         and n.reesculturado),
-    coalesce((select sum(case when ms.tipo = 'SALIDA' then -ms.cantidad else ms.cantidad end)
+    -- ::numeric obligatorio: cantidad es entera, sum() daría bigint y RETURN
+    -- QUERY exige el tipo declarado exacto (42804). Ver el fix
+    -- tyrecontrol_fix_resumen_usados_tipo.sql.
+    coalesce((select sum(case when ms.tipo = 'SALIDA' then -ms.cantidad else ms.cantidad end)::numeric
                 from movimientos_stock ms
                where ms.cliente_id = v_cliente and ms.condicion = 'usado'), 0);
 end $$;
